requirements.txt
pyTelegramBotAPI
python-dotenv
python-telegram-bot
google-generativeai